+++ Product Management/Models/Product.cs
@@ -15,8 +15,8 @@ namespace Product_Management.Models
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
         public int? CategoryId { get; set; }
-        public decimal? PurchasePrice { get; set; }
-        public decimal? SellingPrice { get; set; }
+        public int? PurchasePrice { get; set; }
+        public int? SellingPrice { get; set; }
         public int? Quantity { get; set; }
 
         public virtual Category? Category { get; set; }
